% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{alignK}
\alias{alignK}
\title{Align clusters}
\usage{
alignK(qlist, type = "auto")
}
\arguments{
\item{qlist}{A qlist object}

\item{type}{A character denoting whether clusters must be aligned within K 
or across K. Select 'auto', 'within' or 'across'. Defaults to 'auto'.}
}
\value{
A qlist object with clusters switched where necessary.
}
\description{
Aligns clusters within or across K.
}
\details{
When 'within' is selected, clusters are aligned across replicate runs
within a given K. When 'across' is selected, clusters are aligned across 
replicate runs within a given K and then across K. When 'auto' is selected,
if all runs belong to a single K, clusters are aligned within that K, and if 
multiple K is present, the 'across' method is used.
}
\examples{

sfiles <- list.files(path=system.file("files/structure",package="pophelper"), full.names=TRUE)
slist <- readQ(sfiles)
xlist <- alignK(slist)

}
