% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{analyseRuns}
\alias{analyseRuns}
\title{Analyse STRUCTURE or TESS runs. Wrapper around several smaller functions.}
\usage{
analyseRuns(files = NULL, evannoMethod = TRUE, clumppExport = TRUE,
  plotRuns = TRUE, runsToDf = TRUE, imgoutput = "sep", poplab = NA,
  popcol = NA, writetable = TRUE, sorttable = TRUE, quiet = FALSE)
}
\arguments{
\item{files}{A character or character vector of one or more STRUCTURE or TESS run files}

\item{evannoMethod}{A logical indicating if evanno method should be performed. Applies only to STRUCTURE runs.}

\item{clumppExport}{A logical indicating if files must be exported for clumpp.}

\item{plotRuns}{A logical indicating if selected files should be exported as barplots.}

\item{runsToDf}{A logical indicating if selected files should be converted and returned as an R object (dataframe or a list).}

\item{imgoutput}{A character indicating if files are plotted as separate image files ("sep") or joined into a single image ("join").}

\item{poplab}{A character vector of population labels equal to length of individuals. Each pop name must repeat to the number of individuals present in each pop.}

\item{popcol}{A character vector of colours (representing populations) for colouring clusters. If NA, colours are automatically generated. K 1 to 12 are custom unique colours while K>12 are coloured by function rich.color().}

\item{writetable}{A logical T or F. Setting to TRUE writes the output table to the working directory.}

\item{sorttable}{A logical indicating if the output table must be sorted. Sorts table by loci, ind and K when available.}

\item{quiet}{A logical indicating if messages must be printed}
}
\value{
If a single file is selected, a single dataframe is returned. If
multiple files are selected, a list with multiple dataframes is returned.
}
\description{
A single function to analyse STRUCTURE or TESS runs. Performs tabulate, summarise, evanno method, clumpp export, plots runs and converts runs to an R object.
}
\details{
The function \code{analyseRuns} is a wrapper around several other \code{pophelper} functions. All arguments for all these other functions are not available. If more arguments/options are required, consider running the functions separately.
}

