% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlist.R
\name{as.qlist}
\alias{as.qlist}
\title{Convert a qlist or qlist-like list to a valid qlist.}
\usage{
as.qlist(qlist, debug = FALSE)
}
\arguments{
\item{qlist}{A qlist or a list of runs (data.frames)}

\item{debug}{Logical for internal use}
}
\value{
A qlist object (list of data.frames)
}
\description{
Convert a list to a valid qlist. Adds run labels if needed. 
Renames cluster columns and adds basic attributes 'ind' and 'k' if needed.
}
\examples{

# create two data frames
df1 <- data.frame(A=c(0.2,0.4,0.6,0.2), B=c(0.8,0.6,0.4,0.8))
df2 <- data.frame(A=c(0.3,0.1,0.5,0.6), B=c(0.7,0.9,0.5,0.4))

# create qlist
q1 <- list(df1,df2)
q1

# is.qlist(q1) # error because list elements are not named
q2 <- as.qlist(q1)
is.qlist(q2) # no error
q2

}
