% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{checkQ}
\alias{checkQ}
\title{Internal: Check input filetype.}
\usage{
checkQ(files = NULL, warn = FALSE)
}
\arguments{
\item{files}{A character or character vector of one or more input text files 
or a list of dataframes.}

\item{warn}{A logical indicating if a warning must be displayed for file 
those are not STRUCTURE, TESS or BASIC file.}
}
\value{
A character or character vector indicating input type 'STRUCTURE', 
'TESS', 'BASIC', 'CLUMPP', 'list', 'data.frame', 'UNIDENTIFIED' for all 
selected files.
}
\description{
Internal: Check input filetype.
}
