% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{checkRuns}
\alias{checkRuns}
\title{Internal: Check if a selected run is STRUCTURE or TESS file.}
\usage{
checkRuns(files = NULL, warn = FALSE)
}
\arguments{
\item{files}{A character or character vector of one or more STRUCTURE or TESS run files}

\item{warn}{A logical indicating if a warning be displayed for file those are not STRUCTURE or TESS.}
}
\value{
A character or character vector indicating 'STRUCTURE' or 'TESS' for all selected files.
}
\description{
Check if a selected run is STRUCTURE or TESS file.
}

