% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{clumppExportAdmixture}
\alias{clumppExportAdmixture}
\title{Deprecated}
\usage{
clumppExportAdmixture(files = NULL, prefix = NA, parammode = NA,
  paramrep = NA, useexe = FALSE)
}
\description{
Deprecated
}
\seealso{
\code{\link{clumppExport}}
}

