% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{clumppExportMatrix}
\alias{clumppExportMatrix}
\title{Deprecated}
\usage{
clumppExportMatrix(files = NULL, prefix = NA, parammode = NA,
  paramrep = NA, useexe = FALSE)
}
\description{
Deprecated
}
\seealso{
\code{\link{clumppExport}}
}

