% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{clumppExportStructure}
\alias{clumppExportStructure}
\title{Deprecated}
\usage{
clumppExportStructure(files = NULL, prefix = NA, parammode = NA,
  paramrep = NA, useexe = FALSE)
}
\description{
Deprecated
}
\seealso{
\code{\link{clumppExport}}
}

