% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{collectRunsTess}
\alias{collectRunsTess}
\title{Collect TESS run results from multiple folders}
\usage{
collectRunsTess(runsdir = NULL, newdir = NULL)
}
\arguments{
\item{runsdir}{A character path indicating the directory containing TESS runs in multiple directories. To use current working directory, set \code{runsdir=getwd()}. Use \code{choose.dir()} for interactively selecting the directory.}

\item{newdir}{A character indicating the name/path of the new directory to be
created with the collected runs.}
}
\value{
Two integers are returned. The first denotes the number of TESS run
files copied and renamed. The second number denotes number of directories
without TESS run files.
}
\description{
Collect TESS cluster run files from multiple folders to one
folder and rename each run by folder name
}
\details{
Within each TESS run folder, the function searches for filename
ending with 'TR.txt' as the cluster file. This file is copied to the new
folder and renamed as the name of the respective run directory. Therefore,
DO NOT manually rename original run files or directories.

See the vignette for more details.
}
\examples{
\dontrun{
collectRunsTess(runsdir=getwd(),newdir="output")
}
}
