% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{determineRowsAndCols}
\alias{determineRowsAndCols}
\title{Internal: Determine rows and columns for arbitrary number of plots}
\usage{
determineRowsAndCols(numplots = NA)
}
\arguments{
\item{numplots}{A numeric indicating the number of plots available for plot}
}
\value{
A 2 value vector wih first value denoting row and second value as column.
}
\description{
Internal: Determine rows and columns for figures from arbitrary number of plots
}

