% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{distructColours}
\alias{distructColours}
\alias{distructColors}
\title{Vector of 90 Distruct colours}
\usage{
distructColours()
}
\value{
Returns a character vector of 90 colours recognised by Distruct.
}
\description{
Vector of 90 Distruct colours.
}
\examples{
distructColours();
distructColours()[1:5]; 
distructColours()[10:15];
}
