% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{distructExport}
\alias{distructExport}
\title{Generate input files for DISTRUCT.}
\usage{
distructExport(files = NULL, poplabbottom = NA, poplabtop = NA,
  popmean = FALSE, overwritedirs = FALSE, printtitle = FALSE,
  popcol = NA, grayscale = FALSE, printcolorbrewer = FALSE, sepline = T,
  seplinewidth = 0.2, borderlinewidth = 1.2, indlinewidth = 0.2,
  fontsize = 6, topdist = 5, bottomdist = -7, figheight = 36,
  indwidth = 1, orientation = 0, xorigin = NA, yorigin = NA,
  xscale = 1, yscale = 1, toplabangle = 60, bottomlabangle = 60,
  echodata = TRUE, printdata = FALSE, quiet = FALSE, useexe = FALSE)
}
\arguments{
\item{files}{A character/path or vector of files. Input can be STRUCTURE, TESS, MATRIX run files or TAB files. On windows, use \code{choose.files(multi = TRUE)} for interactive selection.}

\item{poplabbottom}{A character vector of population labels to be plotted below the plot. The vector must be the same length as number of individuals. See details.}

\item{poplabtop}{An optional character vector of population labels to be plotted above the plot. The vector must be the same length as number of individuals. See details.}

\item{popmean}{A logical indicating if individual values are to be plotted (F) or population means are to be plotted (T).}

\item{overwritedirs}{A logical indicating if existing directories must be overwritten (T) automatically or not (F).}

\item{printtitle}{A logical indicating if the filename must be printed as the title on the plot.}

\item{popcol}{A character vector of colours equal to the number of clusters. Note these are not R colours. Use \code{distructColours()} or refer to DISTRUCT manual for colours.
With multiple files, number of colours must equal input file with highest number of clusters.}

\item{grayscale}{A logical indicating if clusters must be shown in grayscale.}

\item{printcolorbrewer}{A logical indicating if the colours provided in \code{popcol} are ColorBrewer colours. See details.}

\item{sepline}{A logical indicating if divider lines must be drawn between populations (T).}

\item{seplinewidth}{A numeric indicating width of sepline.}

\item{borderlinewidth}{A numeric indicating width of border around the plot.}

\item{indlinewidth}{A numeric indicating width of border around individual bars and ticks.}

\item{fontsize}{A numeric indicating font size of population labels.}

\item{topdist}{A numeric indicating distance of top labels from the top edge of the plot.}

\item{bottomdist}{A numeric indicating distance of bottom labels from the bottom edge of the plot. Usually a negative number.}

\item{figheight}{A numeric indicating height of the plot.}

\item{indwidth}{A numeric indicating width of each individual bar. The width of the plot depends on this value.}

\item{orientation}{An integer (0,1,2,3) indicating orientation of the plot. See details.}

\item{xorigin}{A numeric indicating lower left x-coordinate of the plot. See details.}

\item{yorigin}{A numeric indicating lower left y-coordinate of the plot. See details.}

\item{xscale}{A numeric indicating scaling for the x direction.}

\item{yscale}{A numeric indicating scaling for the y direction.}

\item{toplabangle}{A numeric between 0 and 180 indicating angle of top labels.}

\item{bottomlabangle}{A numeric between 0 and 180 indicating angle of bottom labels.}

\item{echodata}{A logical. Not really sure what this does.}

\item{printdata}{A logical indicating if head and tail of data must be shown in display on running DISTRUCT.}

\item{quiet}{A logical TRUE or FALSE. Set to TRUE by default to print verbose statements to screen.}

\item{useexe}{A logical indicating if DISTRUCT executable is automatically run based on system OS (experimental).}
}
\value{
The function does not return anything. The function creates directories for each input file and 
populates it with files necessary to run DISTRUCT. The files are individual q-matrix file (xx-indq.txt),
population q-matrix file (xx-popq.txt), a cluster colour file (xx-colours.txt) and drawparams file. If 
population labels were defined, then (xx-poplab-bottom.txt) or (xx-poplab-top.txt) are also exported. The
DISTRUCT executable is run in this directory to generate an xx.ps file.
}
\description{
Create DISTRUCT input files from STRUCTURE/TESS/MATRIX runs or TAB files.
}
\details{
\strong{Orientation} \cr
0 for horizontal orientation (default) \cr
1 for vertical orientation \cr
2 for reverse horizontal orientation \cr
3 for reverse vertical orientation \cr
\cr
\strong{Origin} \cr
Default values of origin for a given orientation: \cr
orientation, xorigin, yorigin \cr
0,72,288 \cr
1,360,72 \cr
2,540,504 \cr
3,288,720 \cr
If plot exceeds canvas size, consider shifting \code{xorigin} to the left and/or decreasing \code{indwidth}.\cr
\cr
\strong{Colorbrewer colours} \cr
Colorbrewer colours are not automatically generated for now. Refer to DISTRUCT manual for colour names. \cr
Replace the colour names in xx-colours.txt output file with selected colorbrewer colours (ex: Accent_3_qual_1). \cr
\cr
\strong{useexe}\cr
This option automatically runs the DISTRUCT executable that is provided with this package. 
Note that the executable is 64bit LSB requiring a 64bit system. The DISTRUCT executable was obtained from 
here \url{https://web.stanford.edu/group/rosenberglab/distruct.html}. Remember to cite DISTRUCT if this option is used.\cr
}
\examples{
\dontrun{
#read some data
sfiles <- list.files(path=system.file("files/structure",package="pophelper"),full.names=TRUE)
pops <- read.delim(system.file("files/structurepoplabels.txt",package="pophelper"),header=FALSE)

#plot without labels
distructExport(sfiles)

#plot with bottom label
distructExport(files,poplabbottom=pops$V1)

#plot with top label
distructExport(files,poplabtop=pops$V1)

#plot population mean values
distructExport(files,poplabbottom=pops$V1,popmean=T)
}
}

