% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{getDim}
\alias{getDim}
\title{Internal: Get dimensions for figures.}
\usage{
getDim(ind, units = NA, height = NA, width = NA, res = NA,
  plotnum = NA)
}
\arguments{
\item{ind}{A numeric indicating the number of individuals.}

\item{units}{A character indicating the unit of dimension: "cm","mm","in".}

\item{height}{A numeric indicating the height of each plot.}

\item{width}{A numeric indicating the width of each plot.}

\item{res}{A numeric indicating the resolution of the figure.}

\item{plotnum}{A numeric indicating the number of plots in the figure.}
}
\value{
a vector with height and width.
}
\description{
Generate height and width of figure based on number of individuals.
This is an internal function that calculates figure dimensions for export.
}

