% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{getPlotParams}
\alias{getPlotParams}
\title{Internal: Generate parameters for plots with labels}
\usage{
getPlotParams(poplab = NA, plotnum = 1, labpos = NA, labsize = NA,
  labangle = NA, labjust = NA, pointsize = NA, linepos = NA,
  linethick = NA, fmar = NA)
}
\arguments{
\item{poplab}{A character vector of labels same length as number of individuals.}

\item{plotnum}{A numeric indicating the number of plots on the figure.}

\item{labpos}{A numeric indicating the y position of the labels.}

\item{labsize}{A numeric indicating the size of the labels.}

\item{labangle}{A numeric indicating the angle/rotation of labels. 0 is horizontal while 90 is vertical.}

\item{labjust}{A numeric indicating the justification of labels. Defaults to 0.5 if labangle = 0  or 1 if
labangle between 20 and 135.}

\item{pointsize}{A numeric indicating the size of points on label marker line.}

\item{linepos}{A numeric indicating the y position of the label marker line and the points.}

\item{linethick}{A numeric indicating the thickness of the label marker line.}

\item{fmar}{A numeric vector with four elements indicating the figure margins in 'lines' unit. A vector of 4 numbers for top, right,
bottom and left margins. ex. c(0.2,0.2,0.2,0).}
}
\value{
A list with following plot parameters: poplab, plotnum, labpos, labsize,
labangle, labjust, pointsize, linepos, linethick and fmar.
}
\description{
Generates various parameters required for plotting with labels. Internal function.
}

