% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{grpLabels}
\alias{grpLabels}
\title{Internal: Handles grp subset/order}
\usage{
grpLabels(df = NULL, grplab = NA, subsetgrp = NA, grpmean = FALSE,
  labpos = NA, linepos = NA)
}
\arguments{
\item{df}{A q-matrix dataframe}

\item{grplab}{A character vector of group labels}

\item{subsetgrp}{A character or character vector of grp name(s) to subset/reorder}

\item{grpmean}{A logical indicating if q-matrix must be converted from individual to group mean.}

\item{labpos}{A numeric indicating y-axis position of labels}

\item{linepos}{A numeric indicating y-axis position of label line}
}
\value{
Returns a list with subsetted/reordered q-matrix and a character vector 
of original/subsetted/reordered grp label vector. If labpos and linepos is not NA, then 
they are included in the list.
}
\description{
Internal: Takes a q-matrix dataframe along with grp labels. Pop labels
can be reordered or subsetted. The function also creates labelpos and markerpos dfs.
}
