% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlist.R
\name{joinQ}
\alias{joinQ}
\title{Combine two or more qlist objects into one}
\usage{
joinQ(...)
}
\arguments{
\item{...}{Two or more qlist objects separated by comma (,)}
}
\value{
A joined qlist object
}
\description{
Combine two or more qlist objects
}
\examples{

slist <- readQ(list.files(path=system.file("files/structure",package="pophelper"), full.names=TRUE))
tlist <- readQ(list.files(path=system.file("files/tess",package="pophelper"), full.names=TRUE))
jlist <- joinQ(slist,tlist)
length(slist)
length(tlist)
length(jlist)

}
