% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlist.R
\name{mergeQ}
\alias{mergeQ}
\title{Merge replicate runs of the same K}
\usage{
mergeQ(qlist)
}
\arguments{
\item{qlist}{A qlist object (with aligned clusters)}
}
\value{
A merged qlist object. Runs will be ordered by K in the output. Output
runs are labelled by K value.
}
\description{
Merge replicate runs of the same K
}
\details{
Make sure clusters are aligned between replicate runs before merging 
them.
}
\examples{

sfiles <- list.files(path=system.file("files/structure",package="pophelper"), full.names=TRUE)
slist <- readQ(sfiles)
slist <- alignK(slist)
length(slist)
length(mergeQ(slist))
}
\seealso{
\code{\link{alignK}}
}
