% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{plotQMultiline}
\alias{plotQMultiline}
\title{Plot a qlist as individual-level multiline barplot}
\usage{
plotQMultiline(qlist = NULL, spl = NA, lpp = NA, clustercol = NA,
  indlab = TRUE, useindlab = FALSE, sortind = NA, na.rm = FALSE,
  barwidth = 0.9, ticks = FALSE, yaxislabs = FALSE, labsize = 5,
  labangle = 90, labvjust = 0.5, labhjust = 1, imgtype = "png",
  height = NA, width = NA, res = NA, units = NA)
}
\arguments{
\item{qlist}{A qlist (list of dataframes). An output from \code{\link{readQ}}.}

\item{spl}{An integer indicating samples per line. Defaults to 60.}

\item{lpp}{An integer indicating lines per page. Defaults to 11.}

\item{clustercol}{A character vector of colours for clusters.}

\item{indlab}{A logical indicating if individual labels must be shown below the bars. To hide labels, set \code{indlab=FALSE}.}

\item{useindlab}{A logical indicating if individual labels must be read from the rownames of qlist dataframes and used as labels.}

\item{sortind}{A character indicating how individuals are sorted. Default is NA. Other options are 'all' or any one cluster (eg. 'Cluster1').}

\item{na.rm}{Default set to FALSE. NAs are not removed from data, therefore \code{ggplot} prints warning messages for NAs. If set to TRUE, NAs are removed before plotting and \code{ggplot} NA warning is suppressed.}

\item{barwidth}{A numeric indicating the width of the bars. Defaults to 0.9.}

\item{ticks}{A logical indicating if ticks on axis should be displayed or not.}

\item{yaxislabs}{A logical indicating if y-axis labels should be displayed or not.}

\item{labsize}{A numeric denoting size of the labels. Defaults to 5.}

\item{labangle}{A numeric denoting the angle of the labels. Defaults to 90.}

\item{labvjust}{A numeric denoting vertical justification of the labels. Defaults to 0.5.}

\item{labhjust}{A numeric denoting the horizontal justification of the labels. Defaults to 1.}

\item{imgtype}{A character denoting figure output format. Options are 'png', 'jpeg' or 'pdf'. If pdf, height and width must be in inches and res argument is ignored.}

\item{height}{A numeric denoting height of the full figure. If NA, height is set to 29.7cm (A4 height).}

\item{width}{A numeric denoting width of the full figure. If NA, width is set to 21cm (A4 width).}

\item{res}{A numeric denoting resolution of the figure. Default is 300.}

\item{units}{A character denoting the units of dimension of the figure. Default is "cm". Other options can be "in", "mm" or "px".}
}
\description{
Plot a qlist as individual-level barplot with multiple lines.
}
\details{
Figures are always created to A4 size. Any plotted row will span the width of the figure. Note that this function is slow and may take several minutes when plotting mutiple runs.

See the \href{http://royfrancis.github.io/pophelper/}{vignette} for more details.
}
\examples{
\dontrun{
sfiles <- list.files(path=system.file("files/structure",package="pophelper"),full.names=TRUE)
slist <- readQ(sfiles)

# basic
plotQMultiline(slist[1])

# multiple files
plotQMultiline(slist[1:3])

# adjust samples per line and lines per page
plotQMultiline(slist[1],spl=75,lpp=10)

# sort individuals
plotQMultiline(slist[1],sortind="all")
plotQMultiline(slist[1],sortind="Cluster1")

# include labels from file
inds <- read.delim(system.file("files/structureindlabels.txt",package="pophelper"),header=FALSE,stringsAsFactors=FALSE)
rownames(slist[[1]]) <- inds$V1
plotQMultiline(slist[1],useindlab=T)

# change cluster colours
plotQMultiline(slist[1],clustercol=c("steelblue","coral"))

# change bar width and height
plotQMultiline(slist[1],barwidth=1,spl=149,labsize=3,height=5)

}
}

