% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{plotRuns}
\alias{plotRuns}
\title{Plot STRUCTURE, TESS or table files as barplots.}
\usage{
plotRuns(files = NULL, imgoutput = "sep", poplab = NA, popcol = NA,
  sortind = NA, subsetpops = NA, na.rm = TRUE, imgtype = "png",
  height = NA, width = NA, dpi = NA, units = NA, panelspacer = NA,
  flab = TRUE, flabsize = NA, flabcol = NA, flabbackcol = NA,
  labspacer = NA, labpanelheight = NA, labpos = NA, labsize = NA,
  labangle = NA, labjust = NA, labcol = NA, pointsize = NA,
  pointcol = NA, pointbgcol = NA, pointtype = NA, linepos = NA,
  linethick = NA, linecol = NA, linetype = NA, div = TRUE,
  divcol = NA, divtype = NA, divthick = NA)
}
\arguments{
\item{files}{A character vector of one or more STRUCTURE/TESS/table files}

\item{imgoutput}{A character with options: 'sep','join' or'tab'.If set to "sep", STRUCTURE/TESS run files are plotted as separate image files. If set to "join", STRUCTURE/TESS run files are joined into a single image. If set to "tab", combined/aligned/merged files are plotted into separate or joined plots based on number of tables within each file.}

\item{poplab}{A character vector of population labels equal to length of individuals. Each pop name must repeat to the number of individuals present in each pop.}

\item{popcol}{A vector of colours (representing populations) for colouring clusters. If NA, colours are automatically generated. K=1 to 12 are custom unique colours while K>12 are coloured by function rich.color().}

\item{sortind}{A character indicating how individuals are sorted. Default is NA. Other options are 'all' or any one cluster (eg. 'Cluster1').}

\item{subsetpops}{A character or character vector with population names to subset or reorder populations. Only applicable if using 'poplab'. Default is NA. See details.}

\item{na.rm}{A logical indicating if NAs are removed from data, else \code{ggplot} prints warning messages for NAs. If set to TRUE, NAs are removed before plotting and \code{ggplot} NA warning is suppressed.}

\item{imgtype}{A character indicating image file type. Possible options are "png","jpeg" or "pdf". For pdf, height and width are in inches and res does not apply.}

\item{height}{A numeric indicating the height of individual run plot. By default, automatically generated based on number of Individuals. Ranges between 2.5cm and 4.6cm.}

\item{width}{A numeric indicating the width of individual run plot. By default, automatically generated based on number of individuals. Ranges between 5cm and 30cm.}

\item{dpi}{A numeric indicating the image resolution in pixels per inch (PPI). Defaults to 300.}

\item{units}{A numeric indicating the units of height and width. Default set to "cm".}

\item{panelspacer}{A numeric indicating the space at the bottom of one or more barplot panels.}

\item{flab}{A logical indicating if strip panels on right side must be shown. Strip panels display file name and K value. Defaults to TRUE.}

\item{flabsize}{A numeric indicating the size of the filename label. Defaults to 4. Applicable only if flab=T.}

\item{flabcol}{A colour character indicating the colour of the filename label. Defaults to "grey10". Applicable only if flab=T.}

\item{flabbackcol}{A colour character denoting the background colour of the filename label. Defaults to white. Applicable only if flab=T.}

\item{labspacer}{A numeric indicating the space between the plots and the label area. Applicable only with population labels.}

\item{labpanelheight}{A numeric indicating the height of the label area in cm. Default is 0.4 cm. If units are different, cm will be be converted automatically. Applicable only with population labels.}

\item{labpos}{A numeric indicating the y position of the labels. Applicable only with population labels.}

\item{labsize}{A numeric indicating the size of the labels.}

\item{labangle}{A numeric indicating the angle/rotation of labels. 0 is horizontal while 90 is vertical.}

\item{labjust}{A numeric indicating the justification of labels. Defaults to 0.5 if labangle = 0  or 1 if labangle between 20 and 135.}

\item{labcol}{A colour character for the colour of labels. Defaults to "grey30".}

\item{pointsize}{A numeric indicating the size of points on label marker line.}

\item{pointcol}{A colour character for the colour of points on the label marker line. Defaults to "grey30".}

\item{pointbgcol}{A colour character for the background of marker point for certain pointtypes.}

\item{pointtype}{A character or number for the type of points on the label marker line. Defaults to |. Same as pch in standard R.}

\item{linepos}{A numeric indicating the y position of the label marker line and the points.}

\item{linethick}{A numeric indicating the thickness of the label marker line.}

\item{linecol}{A colour character for the label marker line. Defaults to "grey30".}

\item{linetype}{A numeric indicating the type of line for marker line. Same as lty in standard R. Default value is 1.}

\item{div}{A logical indicating if divider lines between population clusters must be drawn. Applicable only with population labels.}

\item{divcol}{A character or hexadecimal colour denoting the colour of the divider line. Default is white.}

\item{divtype}{A numeric indicating the type of line for the divider line. Same as lty in standard R. Default value is 1.}

\item{divthick}{A numeric indicating the thickness of the divider line. Default is 0.25}
}
\value{
Nothing is returned.
}
\description{
Plot one or more STRUCTURE/TESS output files or table files (aligned/combined/merged) files. The STRUCTURE and TESS files can be plotted individually or joined together. The table files will be plotted based on number of runs in each file.
}
\details{
It is possible to set either height or width and leave other as default.\cr
\strong{sortind}\cr
This argument takes one character as input. Individuals can be ordered by one cluster. For ex. \code{sortind="Cluster1"} or \code{sortind="Cluster2"}.
To order by all clusters as the 'Sort by Q' option in STRUCTURE software, use \code{sortind="all"}. When using \code{sortind} with \code{subsetpops}, individuals
are sorted within the population groups.
\strong{subsetpops}\cr
This argument takes one or more characters as input. Use only populations labels used in the poplab vector. For ex. In case of two pops in order 'Pop A' and 'Pop B'.
Use \code{subsetpops=c("Pop B","Pop A")} to change order of populations. Use \code{subsetpops="Pop B"} to subset only Pop B.
}
\seealso{
\code{\link{plotMultiline}}
}

