% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{popLabels}
\alias{popLabels}
\title{Internal: Handles pop subset/order}
\usage{
popLabels(df = NULL, poplab = NA, subsetpops = NA, popmean = FALSE,
  labpos = NA, linepos = NA)
}
\arguments{
\item{df}{A q-matrix dataframe}

\item{poplab}{A character vector of population labels}

\item{subsetpops}{A character or character vector of pop names to subset/reorder}

\item{popmean}{A logical indicating if q-matrix must be converted from individual to population mean.}

\item{labpos}{A numeric indicating y-axis position of labels}

\item{linepos}{A numeric indicating y-axis position of label line}
}
\value{
Returns a list with subsetted/reordered q-matrix and a character vector 
or subsetted/reordered pop label vector. If labpos and linepos is not NA, then they are included in the list.
}
\description{
Internal: Takes a cluster matrix along with pop labels and pop order. If pop order is different,
cluster matrix is reordered by pops without sorting individuals. Also creates labelpos and markerpos dfs.
}

