% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readQBaps}
\alias{readQBaps}
\title{Convert BAPS cluster files to qlist.}
\usage{
readQBaps(files = NULL)
}
\arguments{
\item{files}{A character or character vector of one or more BAPS cluster run 
files. Use \code{choose.files(multi=TRUE)} to select interactively.}
}
\value{
A list of lists with dataframes is returned. List items are named by 
input filename.
}
\description{
Takes one or more BAPS cluster run files and converts them to a 
list of dataframes.
}
\details{
See the \href{http://royfrancis.github.io/pophelper/}{vignette} for 
more details.
}
\examples{
bfiles <- list.files(path=system.file("files/baps",package="pophelper"),
full.names=TRUE)
# create a qlist
blist <- readQBaps(bfiles)
}
