% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{readQBasic}
\alias{readQBasic}
\title{Convert delimited text files to qlist.}
\usage{
readQBasic(files = NULL)
}
\arguments{
\item{files}{A character or character vector of one or more delimited text files. Use \code{choose.files(multi=TRUE)} 
to select interactively.}
}
\value{
A list of lists with dataframes is returned. List items are named by input filename.
}
\description{
Takes one or more delimited numeric text files and converts each of 
them to separate dataframes.
}
\details{
Input files can be Admixture run files, fastStructure meanQ files. 
or any tab-delimited, space-delimited or comma-delimited tabular data without header.
}
\examples{
afiles <- list.files(path=system.file("files/admixture",package="pophelper"),full.names=TRUE)
# create a qlist
readQBasic(afiles)
}
