% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readQClumpp}
\alias{readQClumpp}
\title{Convert CLUMPP format numeric text files to qlist.}
\usage{
readQClumpp(files = NULL)
}
\arguments{
\item{files}{A character or character vector of one or more COMBINED, ALIGNED 
or MERGED files. COMBINED files are generated from \code{clumppExport}. 
ALIGNED and MERGED files are generated by CLUMPP. 
Use \code{choose.files(multi=TRUE)} to select interactively.}
}
\value{
A list of lists with dataframes is returned. Each list item is named 
by input filename. Multiple runs within one file are suffixed by -1, -2 etc.
}
\description{
Takes one or more CLUMPP format numeric text files and converts
them to a list of dataframes.
}
\examples{
cfiles1 <- system.file("files/STRUCTUREpop_K4-combined.txt",package="pophelper")
cfiles2 <- system.file("files/STRUCTUREpop_K4-combined-aligned.txt",
package="pophelper")
cfiles3 <- system.file("files/STRUCTUREpop_K4-combined-merged.txt",
package="pophelper")

# create a qlist
clist1 <- readQClumpp(cfiles1)
clist2 <- readQClumpp(cfiles2)
clist3 <- readQClumpp(cfiles3)

}
