% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readQTess3}
\alias{readQTess3}
\title{Convert TESS3 R object to pophelper qlist.}
\usage{
readQTess3(t3list = NULL, progressbar = FALSE)
}
\arguments{
\item{t3list}{A TESS3 object. An output from function \code{tess3()} from 
package \code{tess3r}. This is not an external object. If imported from outside
of R, it must be read in as an Rdata or Rds file.}

\item{progressbar}{A logical indicating if execution progress must be shown.}
}
\value{
A list of lists with dataframes (qlist) is returned. Each list item 
is named by as sample1, sample2 etc. Within TESS3, 'tess3Main' attributes L, 
n, ploidy, K, rmse and crossentropy are preserved as attributes in the qlist 
dataframe.
}
\description{
Takes a TESS3 R object and convert to pophelper qlist for use 
with pophelper.
}
\details{
See the \href{http://royfrancis.github.io/pophelper/}{vignette} for 
more details.
}
