% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{runsToDfAdmixture}
\alias{runsToDfAdmixture}
\title{Convert ADMIXTURE cluster files to R dataframe.}
\usage{
runsToDfAdmixture(files = NULL)
}
\arguments{
\item{files}{A character or character vector of one or more ADMIXTURE cluster run files. Use \code{choose.files(multi = TRUE)} 
to select interactively.}
}
\value{
If a single file is selected, a single dataframe is returned. 
If multiple files are selected, a list with multiple dataframes is returned.
}
\description{
DEPRECATED: Takes one or more ADMIXTURE cluster run files and converts each of 
them to separate dataframes.
}
\seealso{
\code{\link{runsToDfStructure}}, \code{\link{runsToDfTess}}
}

