% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{runsToDfMatrix}
\alias{runsToDfMatrix}
\title{Convert MATRIX cluster files to R dataframe.}
\usage{
runsToDfMatrix(files = NULL)
}
\arguments{
\item{files}{A character or character vector of one or more ADMIXTURE cluster run files. Use \code{choose.files(multi = TRUE)} 
to select interactively.}
}
\value{
If a single file is selected, a single dataframe is returned. 
If multiple files are selected, a list with multiple dataframes is returned.
}
\description{
Takes one or more MATRIX cluster run files (tabular files, Admixture, fastStructure etc) and converts each of 
them to separate dataframes.
}
\details{
Input files are expected to be Admixture run files or fastStructure meanQ files. 
Input files can also be any tab-delimited, space-delimited or comma-delimited tabular data without header.
}
\seealso{
\code{\link{runsToDfStructure}}, \code{\link{runsToDfTess}}
}

