% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{runsToDfStructure}
\alias{runsToDfStructure}
\title{Convert STRUCTURE run files to R dataframes.}
\usage{
runsToDfStructure(files = NULL, indlabfromfile = FALSE)
}
\arguments{
\item{files}{A character or character vector of one or more STRUCTURE run files. Use \code{choose.files(multi=TRUE)} 
to select interactively.}

\item{indlabfromfile}{A logical indicating if individual labels must be read from input file and used as row names for resulting data frame. Spaces in labels may be replaced with _.}
}
\value{
If a single file is selected, a single dataframe is returned. If 
multiple files are selected, a list with multiple dataframes is returned. 
If individual labels are present in the STRUCTURE file, they are added to the dataframe as row names.
}
\description{
Takes one or more STRUCTURE output files and converts each of 
them to separate dataframes.
}
\examples{
slist <- list.files(path=system.file("files/structure",package="pophelper"),full.names=TRUE)
#create a dataframe
runsToDfStructure(slist[1])

#use ind names from file
runsToDfStructure(slist[1],indlabfromfile=T)

#create a list of dataframes
runsToDfStructure(slist)
}
\seealso{
\code{\link{runsToDfTess}}, \code{\link{runsToDfMatrix}}
}

