% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlist.R
\name{sortQ}
\alias{sortQ}
\title{Sort runs in a qlist based on an attribute}
\usage{
sortQ(qlist, by = "k", decreasing = FALSE, debug = FALSE)
}
\arguments{
\item{qlist}{A qlist object}

\item{by}{A valid attribute name. Generally 'k' or 'ind'. Structure runs can take 
loci, burin, reps etc. Can take more than one name to arrange by multiple fields. For example \code{by=c("ind","k")}.
See details.}

\item{decreasing}{A logical indicating the direction of sorting.}

\item{debug}{Logical for internal use}
}
\value{
A sorted qlist object
}
\description{
Sort runs in a qlist based on an attribute
}
\details{
Argument 'by' can be any attribute name returned by \code{sapply(qlist,attributes)} except 'names', 'class' and 'row.names'. This is usually similar to \code{tabulateQ(qlist,sorttable=FALSE)}.
}
\examples{

sfiles <- list.files(path=system.file("files/structure",package="pophelper"), full.names=TRUE)
slist <- readQ(sfiles)
slist_1 <- sortQ(slist)
names(slist)
names(slist_1)

}
