% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlist.R
\name{splitQ}
\alias{splitQ}
\title{Split a qlist into sublists by attribute}
\usage{
splitQ(qlist, by = "k")
}
\arguments{
\item{qlist}{A qlist object}

\item{by}{A valid attribute name. Generally k or ind. Structure runs can take 
loci, burin, reps etc. Any attribute name returned by \code{sapply(qlist,attributes)}
should work. Defaults to 'k'.}
}
\value{
Returns a list of qlist objects
}
\description{
Split a qlist into sublists by attribute
}
\examples{

sfiles <- list.files(path=system.file("files/structure",package="pophelper"), full.names=TRUE)
slist <- readQ(sfiles)
slist_2 <- splitQ(slist)
str(slist,max.level=0)
str(slist_2,max.level=1)

}
