% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{summariseQ}
\alias{summariseQ}
\alias{summarizeQ}
\title{Summarise a tabulated dataframe}
\usage{
summariseQ(data = NULL, writetable = FALSE, exportpath = NULL)
}
\arguments{
\item{data}{A dataframe with tabulated runs. An output from \code{tabulateQ()}.
Must have minimum 2 columns named k and ind.}

\item{writetable}{A logical indicating if the output table is to be exported
as a tab-delimited text file in the working directory.}

\item{exportpath}{A path to where content must be exported. For example,
\code{exportpath="./dir/anotherdir"}. To use the current working directory, set \code{exportpath=getwd()}.
directory.}
}
\value{
Returns a dataframe with all values of K sorted by K. The table has
3 columns namely value of K, number of runs for each K and number of
individuals.
If the input file is derived from STRUCTURE runs, the table is sorted by loci
as well. Other columns include elpdmean, elpdsd, elpdmin and elpdmax.
}
\description{
Creates a summary table from a tabulated dataframe of two or
more runs with k, number of runs and individuals.
}
\details{
See the \href{http://royfrancis.github.io/pophelper/articles/index.html#summariseq}{vignette} for more details.
}
\examples{

# STRUCTURE files
sfiles <- list.files(path=system.file("files/structure",package="pophelper"),
full.names=TRUE)
slist <- readQ(sfiles)
tr1 <- tabulateQ(slist)
summariseQ(tr1)

# ADMIXTURE files
afiles <- list.files(path=system.file("files/admixture",package="pophelper"),
full.names=TRUE)
tr1 <- tabulateQ(readQ(afiles))
summariseQ(tr1)

}
\seealso{
\code{\link{tabulateQ}}
}
