% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{summariseRunsAdmixture}
\alias{summariseRunsAdmixture}
\title{Deprecated}
\usage{
summariseRunsAdmixture(data = NULL, writetable = FALSE)
}
\description{
Deprecated
}
\seealso{
\code{\link{summariseQ}},
}

