% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{summariseRunsMatrix}
\alias{summariseRunsMatrix}
\alias{summarizeRunsMatrix}
\title{Summarise MATRIX runs}
\usage{
summariseRunsMatrix(data = NULL, writetable = FALSE)
}
\arguments{
\item{data}{A dataframe with tabulated runs. An output from \code{tabulateRunsMatrix()}. Must have minimum 2 columns named k and ind.}

\item{writetable}{A logical indicating if the output table is to be exported as a tab-delimited text file in the working directory.}
}
\value{
Returns a dataframe with all values of K sorted by K. The table has 3 columns namely value of K, number of runs for each K and number of individuals.
}
\description{
Creates a summary table of two or more MATRIX runs with k, number of runs and individuals.
}
\examples{
alist <- list.files(path=system.file("files/admixture",package="pophelper"),full.names=TRUE)
tr1 <- tabulateRunsMatrix(files=alist)
summariseRunsMatrix(tr1)
}
\seealso{
\code{\link{summariseRunsStructure}}, \code{\link{summariseRunsTess}}
}

