% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{summariseRunsStructure}
\alias{summariseRunsStructure}
\title{Deprecated}
\usage{
summariseRunsStructure(data = NULL, writetable = FALSE)
}
\description{
Deprecated.
}
\seealso{
\code{\link{summariseQ}},
}

