% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{tabulateQ}
\alias{tabulateQ}
\title{Tabulate runs from a qlist}
\usage{
tabulateQ(qlist = NULL, writetable = FALSE, sorttable = TRUE)
}
\arguments{
\item{qlist}{A qlist (list of dataframes). An output from \code{\link{readQ}}.}

\item{writetable}{A logical indicating if the output table must be exported as a tab-delimited text file in the working directory.}

\item{sorttable}{A logical indicating if output table is to be sorted. Sorts table by ind and K.}
}
\value{
Returns a dataframe with filenames (if list is not named, then sample1, sample2 etc. is used), K and number of individuals of all runs sorted by ind and K (if \code{sorttable=T}). 
The row numbers of the output table denotes the file number selected. This is helpful if a particular file from the table needs to 
be identified in the selection vector. If input files were derived from STRUCTURE runs, columns loci, burnin, reps, elpd, mvll and vll are also returned.
}
\description{
Takes a qlist of one of more numeric dataframes and creates a table with filenames, K and number of individuals.
}
\details{
The input must be a list of dataframes. If one dataframe is used, then it must be inside a list. If the list items are named, then the item name is used as filename, else sample1, sample2 etc. is used.

See the \href{http://royfrancis.github.io/pophelper/}{vignette} for more details.
}
\examples{

# STRUCTURE files
sfiles <- list.files(path=system.file("files/structure",package="pophelper"),full.names=TRUE)
slist <- readQ(sfiles)
tabulateQ(qlist=slist)

# TESS files
tfiles <- list.files(path=system.file("files/tess",package="pophelper"),full.names=TRUE)
tlist <- readQ(tfiles)
tabulateQ(qlist=tlist)

# ADMIXTURE files
afiles <- list.files(path=system.file("files/admixture",package="pophelper"),full.names=TRUE)
alist <- readQ(afiles)
tabulateQ(qlist=alist)

}
\seealso{
\code{\link{summariseQ}}
}

