% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tabulateRunsTess}
\alias{tabulateRunsTess}
\title{Deprecated}
\usage{
tabulateRunsTess(files = NULL, writetable = FALSE, sorttable = TRUE,
  quiet = FALSE)
}
\description{
Deprecated
}
\seealso{
\code{\link{tabulateQ}},
}

