% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{unitConverter}
\alias{unitConverter}
\title{Internal: Convert between dimension units}
\usage{
unitConverter(value = NA, fromunit = NA, tounit = NA, res = NA)
}
\arguments{
\item{value}{A numeric value or numeric vector to convert}

\item{fromunit}{A character indicating the current unit of the value. Options are "cm", "mm", "in" or "px".}

\item{tounit}{A character indicating the unit to change to. Options are "cm", "mm", "in" or "px".}

\item{res}{A numeric indicating the resolution for pixel conversion. This should be in PPI (pixels per inch).}
}
\value{
Returns a numeric value or numeric vector in changed units.
}
\description{
Internal: Convert value between dimension units
}

