% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{verifyGrplab}
\alias{verifyGrplab}
\title{Verify if a grplab dataframe is formatted correctly.}
\usage{
verifyGrplab(grplab = NULL)
}
\arguments{
\item{grplab}{A dataframe with character fields.}
}
\value{
Nothing.
}
\description{
Verify if a grplab dataframe is formatted correctly.
}
