% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{ellipseCI}
\alias{ellipseCI}
\title{Internal: ellipseCI}
\usage{
ellipseCI(x, y, conf = 0.95, np = 100)
}
\arguments{
\item{x}{A numeric vector of x coordinates}

\item{y}{A numeric vector of y coordinates}

\item{conf}{A numeric indicating confidence interval}

\item{np}{A numeric indicaitng the number of points}
}
\value{
A dataframe with x and y coordinates of the ellipse. Number of
rows is equal to argument np.
}
\description{
Internal: Calculate ellipse for bivariate quantile
}
\details{
Obtained from Claude J (2008) Morphometrics with R, Springer
}

