% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{getColours}
\alias{getColours}
\title{Internal: Get Colours}
\usage{
getColours(k)
}
\arguments{
\item{k}{A numeric indicating the number of colours required}
}
\value{
Returns a character vector of k colours in hexadecimal format
}
\description{
Generate colours based on number of K. This is an internal function to generate colours based on number of populations.
}
\details{
Colours 1 to 12 are custom unique colours. Colours beyond 15 are generated from colour ramp \code{rich.colors()} from package \code{gplots}.
}

