% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{getDim}
\alias{getDim}
\title{Internal: Get dimensions for figures.}
\usage{
getDim(ind = NA, units = NA, height = NA, width = NA, dpi = NA,
  imgtype = NA, labpanelheight = NA, labs = NA, plotnum = NA)
}
\arguments{
\item{ind}{A numeric indicating the number of individuals.}

\item{units}{A character indicating the unit of dimension: "cm","mm","in".}

\item{height}{A numeric indicating the height of each plot.}

\item{width}{A numeric indicating the width of each plot.}

\item{dpi}{A numeric indicating the resolution of the figure.}

\item{imgtype}{A character denoting image format. "png", "jpeg" or "pdf".}

\item{labpanelheight}{A numeric denoting the height of the label panel.}

\item{labs}{An integer denoting number of label groups.}

\item{plotnum}{A numeric indicating the number of plots in the figure.}
}
\value{
a vector with height and width.
}
\description{
Internal: Generate height and width of figure based on number of individuals.
}
