% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{getPlotParams}
\alias{getPlotParams}
\title{Internal: Generate parameters for plots with labels}
\usage{
getPlotParams(poplab = NA, plotnum = 1, labsize = NA, labangle = NA,
  labjust = NA, pointsize = NA, linethick = NA)
}
\arguments{
\item{poplab}{A character vector of labels same length as number of individuals.}

\item{plotnum}{A numeric indicating the number of plots on the figure.}

\item{labsize}{A numeric indicating the size of the labels.}

\item{labangle}{A numeric indicating the angle/rotation of labels. 0 is horizontal while 90 is vertical.}

\item{labjust}{A numeric indicating the justification of labels. Defaults to 0.5 if labangle = 0  or 1 if 
labangle between 20 and 135.}

\item{pointsize}{A numeric indicating the size of points on label marker line.}

\item{linethick}{A numeric indicating the thickness of the label marker line.}
}
\value{
A list with following plot parameters: poplab, plotnum, labsize, 
labangle, labjust, pointsize, linethick.
}
\description{
Internal: Generates various parameters required for plotting with labels.
}

