% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{is.qlist}
\alias{is.qlist}
\title{Verify if a qlist is formatted correctly.}
\usage{
is.qlist(qlist = NULL)
}
\arguments{
\item{qlist}{A qlist object.}
}
\value{
Nothing.
}
\description{
Verify if a qlist is formatted correctly.
}
