% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlist.R
\name{is.qlist}
\alias{is.qlist}
\title{Verify if a qlist is formatted correctly.}
\usage{
is.qlist(qlist = NULL)
}
\arguments{
\item{qlist}{A qlist object.}
}
\value{
Nothing.
}
\description{
Verify if a qlist is formatted correctly.
}
\details{
Returns error if the object is not a list, contents are not data.frames or NA list elements or missing run labels. Warnings are generated for non standard column labels or missing attributes 'ind' and 'k'. Warnings can be fixed by running \code{\link{as.qlist}}.
}
\examples{
sfiles <- list.files(path=system.file("files/structure",package="pophelper"),
full.names=TRUE)
slist <- readQ(sfiles)
is.qlist(slist)
}
\seealso{
\code{\link{as.list}}
}
