% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{readQClumpp}
\alias{readQClumpp}
\title{Convert CLUMPP format numeric text files to qlist.}
\usage{
readQClumpp(files = NULL)
}
\arguments{
\item{files}{A character or character vector of one or more COMBINED, ALIGNED or 
MERGED files. COMBINED files are generated from \code{clumppExport}. ALIGNED and 
MERGED files are generated by CLUMPP. Use \code{choose.files(multi=TRUE)} to 
select interactively.}
}
\value{
A list of lists with dataframes is returned. Each list item is named by 
input filename. Multiple runs within one file are suffixed by -1, -2 etc.
}
\description{
Takes one or more CLUMPP format numeric text files and converts
them to a list of dataframes.
}
\examples{
tabs1 <- system.file("files/STRUCTUREpop_K4-combined.txt",package="pophelper")
tabs2 <- system.file("files/STRUCTUREpop_K4-combined-aligned.txt",package="pophelper")
tabs3 <- system.file("files/STRUCTUREpop_K4-combined-merged.txt",package="pophelper")

#create a qlist
readQClumpp(tabs1)
readQClumpp(tabs2)
readQClumpp(tabs3)

#@export

}
