% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{summariseRunsTess}
\alias{summariseRunsTess}
\title{Summarise TESS runs}
\usage{
summariseRunsTess(data = NULL, writetable = FALSE)
}
\arguments{
\item{data}{A dataframe with tabulated runs. An output from \code{tabulateRunsTess()}. Must have minimum 2 columns named k and ind.}

\item{writetable}{A logical indicating if the output table is to be exported as a tab-delimited text file in the working directory.}
}
\value{
Returns a dataframe with all values of K sorted by K. The table has 3 columns namely value of K, number of runs for each K and number of individuals.
}
\description{
Creates a summary table of several TESS runs with k, number of runs and individuals.
}
\seealso{
\code{\link{summariseRunsStructure}}, \code{\link{summariseRunsMatrix}}
}

