% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pophelper.R
\name{tabulateRunsAdmixture}
\alias{tabulateRunsAdmixture}
\title{Tabulate Admixture runs (Deprecated)}
\usage{
tabulateRunsAdmixture(files = NULL, writetable = FALSE, sorttable = TRUE,
  quiet = FALSE)
}
\arguments{
\item{files}{A character vector of ADMIXTURE cluster files to be tabulated. On windows, use \code{choose.files(multi = TRUE)} for interactive selection.}

\item{writetable}{A logical indicating if output table must be exported as a tab-delimited text file in the working directory.}

\item{sorttable}{A logical indicating if output table is to be sorted. Sorts table by ind and K.}

\item{quiet}{A logical indicating if a message showing the number of selected files is displayed. The message is supressed if set to T.}
}
\value{
Returns a dataframe with filenames, K and number of individuals of all runs sorted by ind and K (if \code{sorttable = T}).
The row numbers of the output table denotes the file number selected. This is helpful if a particular file from the table needs to
be identified in the selection vector.
}
\description{
DEPRECATED: Creates a table from ADMIXTURE output files with filenames, K and number of individuals.
}
\seealso{
\code{\link{tabulateRunsStructure}}, \code{\link{tabulateRunsTess}}
}

