% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{tabulateRunsMatrix}
\alias{tabulateRunsMatrix}
\title{Tabulate Matrix runs}
\usage{
tabulateRunsMatrix(files = NULL, writetable = FALSE, sorttable = TRUE,
  quiet = FALSE)
}
\arguments{
\item{files}{A character vector of MATRIX cluster files to be tabulated. See details. On windows, use \code{choose.files(multi = TRUE)} for interactive selection.}

\item{writetable}{A logical indicating if output table must be exported as a tab-delimited text file in the working directory.}

\item{sorttable}{A logical indicating if output table is to be sorted. Sorts table by ind and K.}

\item{quiet}{A logical indicating if a message showing the number of selected files is displayed. The message is supressed if set to T.}
}
\value{
Returns a dataframe with filenames, K and number of individuals of all runs sorted by ind and K (if \code{sorttable = T}). 
The row numbers of the output table denotes the file number selected. This is helpful if a particular file from the table needs to 
be identified in the selection vector.
}
\description{
Imports tabular data (Admixture, fastStructure etc) and creates a table with filenames, K and number of individuals.
}
\details{
Expected input files are Admixture run files or fastStructure meanQ files. Input files can be any tab-delimited, space-delimited or comma-delimited tabular data without headers.
}
\seealso{
\code{\link{tabulateRunsStructure}}, \code{\link{tabulateRunsTess}}
}

