% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tabulateRunsStructure}
\alias{tabulateRunsStructure}
\title{Deprecated}
\usage{
tabulateRunsStructure(files = NULL, writetable = FALSE, sorttable = TRUE,
  quiet = FALSE)
}
\description{
Deprecated
}
\seealso{
\code{\link{tabulateQ}},
}

