% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pophelper.R
\name{tabulateRunsStructure}
\alias{tabulateRunsStructure}
\title{Tabulate STRUCTURE runs}
\usage{
tabulateRunsStructure(files = NULL, writetable = FALSE, sorttable = TRUE,
  quiet = FALSE)
}
\arguments{
\item{files}{A character or vector of STRUCTURE output files to be tabulated. On windows, use \code{choose.files(multi = TRUE)} for interactive selection.}

\item{writetable}{A logical TRUE or FALSE. Set to FALSE by default. Setting to TRUE writes the output table to the working directory.}

\item{sorttable}{A logical indicating if the output table needs to be sorted. Default set to TRUE. Sorts table by loci, ind and K.}

\item{quiet}{A logical TRUE or FALSE. Set to FALSE by default to print number of selected files. If set to TRUE, then number of selected files are not printed.}
}
\value{
Returns a dataframe with all runs sorted by loci, ind and K (if sorttable = T). The table has 10 columns namely file name, value of K, number of individuals, number of loci, estimated ln probability of data, mean value of ln likelihood, variance of ln likelihood, mean value of alpha, number of burn-in and number of repeats. Missing values are given NA.
The row numbers of the output table denotes the file number selected. This is helpful if a particular file from the table needs to 
be identified in the selection vector.
}
\description{
Creates a table from STRUCTURE output files with various STRUCTURE parameters.
}
\examples{
slist <- list.files(path=system.file("files/structure",package="pophelper"),full.names=TRUE)
tabulateRunsStructure(files=slist)
}
\seealso{
\code{\link{tabulateRunsTess}}, \code{\link{tabulateRunsMatrix}}
}

